feat: fully implemented Delete function into Admin Page + refactor: deleted DeleteItemFragment and adjusted files accordingly + refactor: returned AddItem's xml to original state
