+++ app/src/main/AndroidManifest.xml
@@ -16,9 +16,6 @@
             android:name=".AddItemFragment"
             android:exported="false" />
         <activity
-            android:name=".DeleteItemFragment"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
