+++ app/src/main/java/com/taam/collection_management_system/ManageItemsFragment.java
@@ -27,8 +27,6 @@ public class ManageItemsFragment extends Fragment {
             }
         });
 
-        buttonDeleteItem.setOnClickListener(v -> loadFragment(new DeleteItemFragment()));
-
         buttonBack.setOnClickListener(v -> getParentFragmentManager().popBackStack());
 
         return view;
