+++ app/src/main/res/layout/fragment_add_item.xml
@@ -19,7 +19,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_below="@+id/toolbar"
-        android:layout_marginTop="808dp"
+        android:layout_marginTop="861dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -27,7 +27,7 @@
         android:id="@+id/editTextName"
         android:layout_width="597dp"
         android:layout_height="69dp"
-        android:layout_marginTop="112dp"
+        android:layout_marginTop="108dp"
         android:hint="Name"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -45,7 +45,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/editTextName" />
 
-
     <EditText
         android:id="@+id/editTextDescription"
         android:layout_width="446dp"
@@ -53,15 +52,16 @@
         android:layout_marginTop="4dp"
         android:hint="Description"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/editTextLot" />
+
 
     <Spinner
         android:id="@+id/spinnerCategory"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="48dp"
+        android:layout_marginTop="76dp"
         android:entries="@array/categories_array"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -72,21 +72,9 @@
         android:id="@+id/buttonAdd"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="404dp"
+        android:layout_marginTop="440dp"
         android:text="@string/add_item"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.248"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/spinnerCategory" />
-
-    <Button
-        android:id="@+id/buttonDelete"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="404dp"
-        android:text="Delete Item"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.736"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/spinnerCategory" />
 
@@ -94,10 +82,10 @@
         android:id="@+id/spinnerPeriod"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="52dp"
+        android:layout_marginTop="68dp"
         android:entries="@array/period_array"
         app:layout_constraintTop_toBottomOf="@+id/spinnerCategory"
-        tools:layout_editor_absoluteX="2dp" />
+        tools:layout_editor_absoluteX="0dp" />
 
     <ImageView
         android:id="@+id/imageView"
@@ -112,8 +100,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@id/imageView"
-        android:layout_marginStart="136dp"
-        android:layout_marginTop="512dp"
+        android:layout_marginStart="137dp"
+        android:layout_marginTop="554dp"
         android:text="Select Image/Video"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -122,7 +110,7 @@
         android:id="@+id/editTextText2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="420dp"
+        android:layout_marginTop="456dp"
         android:ems="10"
         android:inputType="text"
         android:text="Choose Period Below"
@@ -131,9 +119,9 @@
 
     <EditText
         android:id="@+id/editTextText4"
-        android:layout_width="213dp"
-        android:layout_height="39dp"
-        android:layout_marginTop="356dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="364dp"
         android:ems="10"
         android:inputType="text"
         android:text="Choose Category Below"
@@ -147,7 +135,7 @@
         android:layout_marginTop="66dp"
         android:ems="10"
         android:inputType="text"
-        android:text="Add a Record Page - To delete a record: press Delete"
+        android:text="To add a new record, please fill out all fields below!"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
