+++ app/src/main/res/layout/fragment_delete_item.xml
@@ -5,8 +5,7 @@
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     app:cardCornerRadius="16dp"
-    app:cardElevation="10dp"
-    tools:context=".DeleteItemFragment">
+    app:cardElevation="10dp" >
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -16,8 +15,8 @@
             android:id="@+id/defaultText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Confirm Delete"
             android:layout_marginTop="25dp"
+            android:text="Confirm Delete"
             android:textSize="30dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -26,32 +25,35 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Are you sure you want to delete LOT NUMBER?"
+            android:text="Are you sure you want to delete the following records?"
             android:textSize="16dp"
-            app:layout_constraintStart_toStartOf="@id/defaultText"
-            app:layout_constraintEnd_toEndOf="@id/defaultText"
-            app:layout_constraintTop_toBottomOf="@id/defaultText"/>
+            android:gravity="center_horizontal"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/defaultText" />
 
-        <EditText
-            android:id="@+id/editLotNumber"
+        <TextView
+            android:id="@+id/selectedLots"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="104dp"
-            android:hint="Lot#"
+            android:text=""
+            android:gravity="center_horizontal"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintHorizontal_bias="0.5"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
         <LinearLayout
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginTop="32dp"
+            android:layout_marginBottom="32dp"
             android:orientation="horizontal"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginTop="32dp"
-            android:layout_marginBottom="32dp"
-            app:layout_constraintTop_toBottomOf="@+id/editLotNumber"
+            app:layout_constraintTop_toBottomOf="@+id/selectedLots"
             tools:layout_editor_absoluteX="112dp"
             tools:layout_editor_absoluteY="163dp">
